Move logger + exceptions to db.utils
