+++ api/utils/gcp.py
@@ -4,44 +4,12 @@ import logging
 
 from google.auth.exceptions import DefaultCredentialsError
 
-levels_map = {'DEBUG': logging.DEBUG, 'INFO': logging.INFO, 'WARNING': logging.WARNING}
 
 IGNORE_GCP_CREDENTIALS_ERROR = os.getenv('SM_IGNORE_GCP_CREDENTIALS_ERROR') in (
     'y',
     'true',
     '1',
 )
-USE_GCP_LOGGING = os.getenv('SM_ENABLE_GCP_LOGGING', '0').lower() in ('y', 'true', '1')
-LOGGING_LEVEL = levels_map[os.getenv('SM_LOGGING_LEVEL', 'DEBUG').upper()]
-
-
-def setup_gcp_logging_if_required(
-    ignore_gcp_credentials_error=IGNORE_GCP_CREDENTIALS_ERROR,
-):
-    """
-    Retrieves a Cloud Logging handler based on the environment
-    you're running in and integrates the handler with the
-    Python logging module. By default this captures all logs
-    at INFO level and higher
-    """
-    logging.basicConfig(level=LOGGING_LEVEL)
-    logger = logging.getLogger('sample-metadata-api')
-
-    if USE_GCP_LOGGING:
-        try:
-            # pylint: disable=import-outside-toplevel,c-extension-no-member
-            import google.cloud.logging
-
-            client = google.cloud.logging.Client()
-            client.get_default_handler()
-            client.setup_logging()
-
-        except DefaultCredentialsError as exp:
-            if not ignore_gcp_credentials_error:
-                raise exp
-
-    return logger
-
 
 # 2021-06-02 mfranklin:
 #   Sometimes it's useful to start the server without a GCP context,
